feat!: don't depend on base64 and serde directly
Because the macro is expanded on the call-site, we can remove the direct
dependencies and have the crate using us as a dep depend on `serde` and
`base64` directly.

As long as the `base64` API does not change significatnly, we should be
fine.

We bump the version as this is backwards incompatible